tests: move package_list tests into a module
