+++ tests/functional/clients/standalone/package_list/fixtures/npms.py
