+++ tests/functional/clients/standalone/package_list/fixtures/pypkgs.py
